LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/unreachable-snap-positions-004-expected.txt



FAIL Snaps to the positions defined by the element as much as possible assert_equals: expected 10 but got 0
